// src/enums/candidate.enum.ts
export enum JobStatus {
  AVAILABLE = 'Available',
  INTERVIEW = 'Interview',
  HIRED = 'Hired',
}

export enum Gender {
  MALE = 'Male',
  FEMALE = 'Female',
}

export enum VerificationStatus {
  IN_PROGRESS = 'In Progress',
  VERIFIED = 'Verified',
  NOT_VERIFIED = 'Not Verified',
}

export enum ProfileStatus {
  COMPLETE = 'Complete',
  IN_COMPLETE = 'Incomplete',
}

export enum Name {
  ASCENDING = 'name',
  DESCENDING = '-name',
}

export enum CreatedAt {
  LATEST = 'createdAt',
  OLDEST = '-createdAt',
}

export enum ProfileMode {
  ACTIVE = 'Active',
  INACTIVE = 'Inactive',
}